[MCUX-69261] examples: Kconfig: ensure task additions headers are pulled in
Due to the fact that both K32W1 as well as RW612 platforms have
configINCLUDE_FREERTOS_TASK_C_ADDITIONS_H set to 1, for the KEX ZIP
generation it's neccessary that the corresponding "task_additions.h"
is packed into the SDK image. As such, we will pull unconditionally
the middleware.freertos-kernel.extension component to ensure that both
the ARMGCC as well as the KEX ZIP package can build succesfully.

Signed-off-by: Alex Porosanu <alexandru.porosanu@nxp.com>