[MCZB-2532][MCZB-2533][MCZB-2534] platform: RW612: override NVIC_SystemReset
Depending on the platform, NVIC_SystemReset might not be enough to
correctly reset the device.
On RW612, NVIC_SystemReset won't reset the core running the MAC.
This is causing some failures during some certification tests.

To fix this, we can override the `NVIC_SystemReset` API.
CMSIS allows to do this by defining `CMSIS_NVIC_VIRTUAL`, it will
include a header defined by the user, which overrides some NVIC macros.

Signed-off-by: Axel Le Bourhis <axel.lebourhis@nxp.com>