[MCZB-2584] examples: Perform factory reset for successful Leave request
User requested factory reset (through hold down the USER button and
press the RESET button) and device started to ZDO Leave procedure. If the
parent is no longer reachable, the leave will be considered as Failed by the
application (return code 0xC2) and the factory reset performed. If the
device still belongs to a network, it will successfully send a Leave frame
and the application should perform the factory reset - as initially requested
by user.

Signed-off-by: Anca Jeanina Dragusin <ancajeanina.dragusin@nxp.com>