[MCZB-2683] ZCL: GreenPower: Fix bIsActAsTempMaster for Commissioning frame
Our Combo Basic Device currently supports a minimum implementation in terms
of tempMaster: Sink is always the tempMaster and no tempMaster election is
implemented.

As such, force the local bIsActAsTempMaster to be always TRUE. In the current
implementation, bIsActAsTempMaster is a member of type bool_t (1 byte)
in union uMessage of tsGP_GreenPowerCallBackMessage structure. For frames of
type Commissioning, the uMessage union is changed to hold a pointer to
tsGP_ZgpCommissionIndication, causing the previous stored value of
bIsActAsTempMaster to be rewritten.

Signed-off-by: Anca Jeanina Dragusin <ancajeanina.dragusin@nxp.com>